+++ Changelog.md
@@ -17,6 +17,8 @@ Version 1.4
 * Menu item for auto-Renumbering of files.
 * Can drag a row ontop of another to copy tags.
 * Genre and Grouping preferences.
+	Defaults genre can be selected
+	auto complete
 
 ### Improvements ###
 * Macruby loading.